Create shared multiplexer_base
